+++ pmd-core/src/main/java/net/sourceforge/pmd/cpd/Tokens.java
@@ -14,7 +14,7 @@ import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import net.sourceforge.pmd.annotation.InternalApi;
-import net.sourceforge.pmd.lang.ast.TokenMgrError;
+import net.sourceforge.pmd.lang.ast.LexException;
 import net.sourceforge.pmd.lang.document.FileId;
 import net.sourceforge.pmd.lang.document.TextDocument;
 
@@ -117,8 +117,8 @@ public class Tokens {
             }
 
             @Override
-            public TokenMgrError makeLexException(int line, int column, String message, @Nullable Throwable cause) {
-                return new TokenMgrError(line, column, fileId, message, cause);
+            public LexException makeLexException(int line, int column, String message, @Nullable Throwable cause) {
+                return new LexException(line, column, fileId, message, cause);
             }
 
             @Override