+++ pmd-core/src/test/java/net/sourceforge/pmd/lang/ast/LexExceptionTest.java
@@ -8,30 +8,30 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
-class TokenMgrErrorTest {
+class LexExceptionTest {
     @Test
     void invalidLocation() {
-        TokenMgrError error = new TokenMgrError(2, 0, null, "test", null);
+        LexException error = new LexException(2, 0, null, "test", null);
         // this shouldn't throw a IllegalArgumentException
         assertEquals("line 2, column 1", error.location().startPosToString());
     }
 
     @Test
     void invalidLocationJavaCC() {
-        TokenMgrError error = new TokenMgrError(false, "DEFAULT", 2, 0, "}", '\n');
+        LexException error = new LexException(false, "DEFAULT", 2, 0, "}", '\n');
         // this shouldn't throw a IllegalArgumentException
         assertEquals("line 2, column 1", error.location().startPosToString());
     }
 
     @Test
     void validLocation() {
-        TokenMgrError error = new TokenMgrError(1, 1, null, "test", null);
+        LexException error = new LexException(1, 1, null, "test", null);
         assertEquals("line 1, column 1", error.location().startPosToString());
     }
 
     @Test
     void validLocationJavaCC() {
-        TokenMgrError error = new TokenMgrError(false, "DEFAULT", 1, 1, "}", '\n');
+        LexException error = new LexException(false, "DEFAULT", 1, 1, "}", '\n');
         assertEquals("line 1, column 1", error.location().startPosToString());
     }
 }