+++ pmd-lang-test/src/main/kotlin/net/sourceforge/pmd/cpd/test/CpdTextComparisonTest.kt
@@ -6,16 +6,13 @@ package net.sourceforge.pmd.cpd.test
 
 import io.kotest.assertions.throwables.shouldThrow
 import net.sourceforge.pmd.cpd.*
-import net.sourceforge.pmd.lang.Language
 import net.sourceforge.pmd.lang.LanguagePropertyBundle
 import net.sourceforge.pmd.lang.LanguageRegistry
-import net.sourceforge.pmd.lang.ast.TokenMgrError
+import net.sourceforge.pmd.lang.ast.LexException
 import net.sourceforge.pmd.lang.document.TextDocument
-import net.sourceforge.pmd.lang.document.TextFile
 import net.sourceforge.pmd.lang.document.FileId
 import net.sourceforge.pmd.test.BaseTextComparisonTest
 import org.apache.commons.lang3.StringUtils
-import java.util.*
 
 /**
  * CPD test comparing a dump of a file against a saved baseline.
@@ -73,18 +70,18 @@ abstract class CpdTextComparisonTest(
     }
 
     @JvmOverloads
-    fun expectTokenMgrError(
+    fun expectLexException(
         source: String,
         fileName: FileId = FileId.UNKNOWN,
         properties: LanguagePropertyConfig = defaultProperties()
-    ): TokenMgrError =
-        expectTokenMgrError(FileData(fileName, source), properties)
+    ): LexException =
+        expectLexException(FileData(fileName, source), properties)
 
     @JvmOverloads
-    fun expectTokenMgrError(
+    fun expectLexException(
         fileData: FileData,
         config: LanguagePropertyConfig = defaultProperties()
-    ): TokenMgrError =
+    ): LexException =
         shouldThrow {
             tokenize(newTokenizer(config), fileData)
         }